Added cases to TransitionTest.
Currently default transition behavior (-> (Target.Kind.EXTERNAL)) is not actually external. It's more like a local transition that changes to external if the target is not a descendant of the source. I think I'll have to make the alternate transition operator (currently >) force an external transition rather than a local one.